+++ packages/cursorless-engine/scripts/transformRecordedTests/transformations/identity.ts
@@ -1,4 +1,4 @@
-import { TestCaseFixture } from "../../../testCaseRecorder/TestCaseFixture";
+import { TestCaseFixture } from "@cursorless/common";
 
 export function identity(fixture: TestCaseFixture) {
   return fixture;
