+++ packages/cursorless-engine/scripts/transformRecordedTests/transformations/upgrade.ts
@@ -1,6 +1,6 @@
 import { flow } from "lodash";
 import { canonicalizeAndValidateCommand } from "../../../core/commandVersionUpgrades/canonicalizeAndValidateCommand";
-import { TestCaseFixture } from "../../../testCaseRecorder/TestCaseFixture";
+import { TestCaseFixture } from "@cursorless/common";
 import { cleanUpTestCaseCommand } from "../../../testUtil/cleanUpTestCaseCommand";
 import { reorderFields } from "./reorderFields";
