+++ packages/cursorless-engine/scripts/transformRecordedTests/transformations/upgradeFromVersion0.ts
@@ -1,5 +1,5 @@
 import { PartialPrimitiveTargetDescriptor } from "@cursorless/common";
-import { TestCaseFixture } from "../../../testCaseRecorder/TestCaseFixture";
+import { TestCaseFixture } from "@cursorless/common";
 import { transformPartialPrimitiveTargets } from "../../../util/getPrimitiveTargets";
 
 export function upgradeFromVersion0(fixture: TestCaseFixture) {
