+++ packages/cursorless-vscode/extension.ts
@@ -1,12 +1,9 @@
 import {
-  Command,
-  CURSORLESS_COMMAND_ID,
   FakeIDE,
   getFakeCommandServerApi,
   isTesting,
   NormalizedIDE,
   Range,
-  TargetPlainObject,
   TextDocument,
 } from "@cursorless/common";
 import {
@@ -14,23 +11,14 @@ import {
   FactoryMap,
   Graph,
   graphFactories,
-  ide,
   injectIde,
   makeGraph,
-  plainObjectToTarget,
-  showCheatsheet,
-  takeSnapshot,
-  TestCaseRecorder,
   ThatMark,
-  updateDefaults,
 } from "@cursorless/cursorless-engine";
 import {
   commandIds,
   KeyboardCommands,
-  showDocumentation,
-  showQuickPick,
   StatusBarItem,
-  toVscodeEditor,
   VscodeIDE,
 } from "@cursorless/cursorless-vscode-core";
 import {
@@ -40,6 +28,8 @@ import {
   toVscodeRange,
 } from "@cursorless/vscode-common";
 import * as vscode from "vscode";
+import { constructTestHelpers } from "./constructTestHelpers";
+import { registerCommands } from "./registerCommands";
 
 /**
  * Extension entrypoint called by VSCode on Cursorless startup.
@@ -105,40 +95,13 @@ export async function activate(
 
   return {
     testHelpers: isTesting()
-      ? {
+      ? constructTestHelpers(
           commandServerApi,
-          ide: ide() as NormalizedIDE,
-          injectIde,
-
-          toVscodeEditor,
-
-          // FIXME: Remove this once we have a better way to get this function
-          // accessible from our tests
-          takeSnapshot,
-
-          setThatMark(
-            editor: vscode.TextEditor,
-            targets: TargetPlainObject[] | undefined,
-          ): void {
-            thatMark.set(
-              targets?.map((target) =>
-                plainObjectToTarget(vscodeIDE.fromVscodeEditor(editor), target),
-              ),
-            );
-          },
-          setSourceMark(
-            editor: vscode.TextEditor,
-            targets: TargetPlainObject[] | undefined,
-          ): void {
-            sourceMark.set(
-              targets?.map((target) =>
-                plainObjectToTarget(vscodeIDE.fromVscodeEditor(editor), target),
-              ),
-            );
-          },
-
-          hatTokenMap: graph.hatTokenMap,
-        }
+          thatMark,
+          sourceMark,
+          vscodeIDE,
+          graph,
+        )
       : undefined,
 
     experimental: {
@@ -150,85 +113,4 @@ export async function activate(
 // this method is called when your extension is deactivated
 export function deactivate() {
   // do nothing
-}
-
-function registerCommands(
-  extensionContext: vscode.ExtensionContext,
-  vscodeIde: VscodeIDE,
-  commandRunner: CommandRunner,
-  testCaseRecorder: TestCaseRecorder,
-  keyboardCommands: KeyboardCommands,
-): void {
-  const commands = [
-    // The core Cursorless command
-    [
-      CURSORLESS_COMMAND_ID,
-      async (spokenFormOrCommand: string | Command, ...rest: unknown[]) => {
-        try {
-          return await commandRunner.runCommandBackwardCompatible(
-            spokenFormOrCommand,
-            ...rest,
-          );
-        } catch (e) {
-          if (!isTesting()) {
-            vscodeIde.handleCommandError(e as Error);
-          }
-          throw e;
-        }
-      },
-    ],
-
-    // Cheatsheet commands
-    ["cursorless.showCheatsheet", showCheatsheet],
-    ["cursorless.internal.updateCheatsheetDefaults", updateDefaults],
-
-    // Testcase recorder commands
-    ["cursorless.recordTestCase", testCaseRecorder.toggle],
-    ["cursorless.pauseRecording", testCaseRecorder.pause],
-    ["cursorless.resumeRecording", testCaseRecorder.resume],
-    ["cursorless.takeSnapshot", testCaseRecorder.takeSnapshot],
-
-    // Other commands
-    [commandIds.showQuickPick, showQuickPick],
-    ["cursorless.showDocumentation", showDocumentation],
-
-    // General keyboard commands
-    [
-      "cursorless.keyboard.escape",
-      keyboardCommands.keyboardHandler.cancelActiveListener,
-    ],
-
-    // Targeted keyboard commands
-    [
-      "cursorless.keyboard.targeted.targetHat",
-      keyboardCommands.targeted.targetDecoratedMark,
-    ],
-    [
-      "cursorless.keyboard.targeted.targetScope",
-      keyboardCommands.targeted.targetScopeType,
-    ],
-    [
-      "cursorless.keyboard.targeted.targetSelection",
-      keyboardCommands.targeted.targetSelection,
-    ],
-    [
-      "cursorless.keyboard.targeted.clearTarget",
-      keyboardCommands.targeted.clearTarget,
-    ],
-    [
-      "cursorless.keyboard.targeted.runActionOnTarget",
-      keyboardCommands.targeted.performActionOnTarget,
-    ],
-
-    // Modal keyboard commands
-    ["cursorless.keyboard.modal.modeOn", keyboardCommands.modal.modeOn],
-    ["cursorless.keyboard.modal.modeOff", keyboardCommands.modal.modeOff],
-    ["cursorless.keyboard.modal.modeToggle", keyboardCommands.modal.modeToggle],
-  ] as const;
-
-  extensionContext.subscriptions.push(
-    ...commands.map(([commandId, callback]) =>
-      vscode.commands.registerCommand(commandId, callback),
-    ),
-  );
 }
